Implement more core operations and documentation tests (#570)
* Implement more operations for std::cmp::Ordering and document.
* Add as operator and document more of std::iter module.
* Add documentation for std::i64 and std::f64.
* Document std::object.
* Start running captured doc comments.
* Move numerical types and implement suffixes.

+++ benches/benches/benchmarks/aoc_2020_1a.rs
@@ -31,7 +31,7 @@ fn aoc_2020_1a(b: &mut Criterion) {
         struct NoSolution;
 
         fn part1(v, target) {
-            v.sort::<int>();
+            v.sort::<i64>();
 
             let a = 0;
             let b = v.len() - 1;
@@ -48,7 +48,7 @@ fn aoc_2020_1a(b: &mut Criterion) {
         }
 
         fn part2(v, target) {
-            v.sort::<int>();
+            v.sort::<i64>();
 
             let a = 0;
             let c = v.len() - 1;
